Add network layer
OkHttpFactory consolidates all settings for the OkHttpClient instance to
be used throughout the app.

ApiService is the Retrofit interface class that defines the API calls
available to the rest of the app. Makes use of rxJava to simplify
threading, as well as get access to fun things like flatmaps.

ApiServiceFactory creates the Retrofit instance used to get a handle on
the ApiService instance used for making API calls.